feat: Add comprehensive test suite for HTTP request execution and response validation
- Introduced multiple test files to validate the execution of HTTP requests from `.http` files, covering various scenarios including single and multiple requests, invalid methods, and custom variables.
- Implemented edge case tests to ensure robustness against empty blocks and invalid configurations.
- Enhanced response validation tests to include checks for headers, body content, and status codes, ensuring accurate comparison against expected responses.
- Added utility functions for creating mock servers and processing test files, streamlining the testing process.
