+++ testdata/http_response_files/validator_status_code_only.hresp
@@ -1,0 +1,1 @@
+HTTP/1.1 200
