--- validator_test.go
@@ -1,1042 +1,0 @@
-package restclient
-
-import (
-	"fmt"
-	"net/http"
-	"strings"
-	"testing"
-
-	"github.com/hashicorp/go-multierror"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-// Helper to return a pointer to an int
-func ptr(i int) *int {
-	return &i
-}
-
-// assertMultierrorContains checks if a multierror.Error contains the expected number of errors
-// and if each of the expected error substrings is present in one of the actual errors.
-func assertMultierrorContains(t *testing.T, err error, expectedErrCount int, expectedErrTexts []string) {
-	t.Helper()
-	require.Error(t, err, "Expected an error, but got nil")
-
-	merr, ok := err.(*multierror.Error)
-	require.True(t, ok, "Expected a *multierror.Error, but got %T", err)
-
-	assert.Len(t, merr.Errors, expectedErrCount, "Mismatch in expected number of errors")
-
-	if len(expectedErrTexts) > 0 {
-		for _, expectedText := range expectedErrTexts {
-			found := false
-			for _, actualErr := range merr.Errors {
-				if strings.Contains(actualErr.Error(), expectedText) {
-					found = true
-					break
-				}
-			}
-			assert.True(t, found, "Expected error text '%s' not found in %v", expectedText, merr.Errors)
-		}
-	}
-}
-
-func TestValidateResponses_NilAndEmptyActuals(t *testing.T) {
-	testFilePath := "testdata/http_response_files/validator_nil_empty_actuals_expected.hresp"
-
-	t.Run("nil actual response slice", func(t *testing.T) {
-		var nilActuals []*Response // nil slice
-		err := ValidateResponses(testFilePath, nilActuals...)
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "mismatch in number of responses: got 0 actual, but expected 1")
-	})
-
-	t.Run("empty actual response slice", func(t *testing.T) {
-		emptyActuals := []*Response{} // empty slice
-		err := ValidateResponses(testFilePath, emptyActuals...)
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "mismatch in number of responses: got 0 actual, but expected 1")
-	})
-
-	t.Run("slice with one nil actual response", func(t *testing.T) {
-		oneNilActual := []*Response{nil}
-		err := ValidateResponses(testFilePath, oneNilActual...)
-		require.Error(t, err)
-		merr, ok := err.(*multierror.Error)
-		require.True(t, ok, "Expected a multierror.Error")
-		require.Len(t, merr.Errors, 1)
-		assert.Contains(t, merr.Errors[0].Error(), "mismatch in number of responses: got 0 actual, but expected 1")
-	})
-}
-
-func TestValidateResponses_FileErrors(t *testing.T) {
-	actualResp := &Response{StatusCode: 200}
-
-	t.Run("missing expected response file", func(t *testing.T) {
-		err := ValidateResponses("nonexistent.hresp", actualResp)
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "failed to parse expected response file")
-		assert.Contains(t, err.Error(), "nonexistent.hresp")
-	})
-
-	t.Run("empty expected response file", func(t *testing.T) {
-		emptyFilePath := "testdata/http_response_files/validator_empty_expected.hresp"
-		err := ValidateResponses(emptyFilePath, actualResp)
-		assertMultierrorContains(t, err, 1, []string{
-			"mismatch in number of responses: got 1 actual, but expected 0",
-		})
-	})
-
-	t.Run("malformed expected response file", func(t *testing.T) {
-		malformedFilePath := "testdata/http_response_files/validator_malformed_status.hresp"
-		err := ValidateResponses(malformedFilePath, actualResp)
-		assertMultierrorContains(t, err, 2, []string{
-			"failed to parse expected response file",
-			"invalid status code",
-			"mismatch in number of responses: got 1 actual, but expected 0",
-		})
-	})
-}
-
-func TestValidateResponses_StatusString(t *testing.T) {
-	tests := []struct {
-		name             string
-		actualResponse   *Response
-		expectedFilePath string
-		expectedErrCount int
-		expectedErrText  string // Kept for single error cases not using multierror helper
-		expectedErrTexts []string
-	}{
-		{
-			name:             "matching status string",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK"},
-			expectedFilePath: "testdata/http_response_files/validator_body_exact_no_body_exp.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "mismatching status string",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 Something Else"},
-			expectedFilePath: "testdata/http_response_files/validator_body_exact_no_body_exp.hresp",
-			expectedErrCount: 1,
-			expectedErrText:  "status string mismatch: expected '200 OK', got '200 Something Else'",
-		},
-		{
-			name:             "actual status string is correct, expected file has only status code",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK"},
-			expectedFilePath: "testdata/http_response_files/validator_partial_status_code_mismatch.hresp",
-			expectedErrCount: 1,
-			expectedErrText:  "status string mismatch: expected '200', got '200 OK'",
-		},
-		{
-			name:             "mismatching status code, status strings also mismatch",
-			actualResponse:   &Response{StatusCode: 404, Status: "404 Not Found"},
-			expectedFilePath: "testdata/http_response_files/validator_body_exact_no_body_exp.hresp",
-			expectedErrCount: 2,
-			expectedErrTexts: []string{"status code mismatch: expected 200, got 404", "status string mismatch: expected '200 OK', got '404 Not Found'"},
-		},
-		{
-			name:             "matching status code, expected file only code, actual also only code in status",
-			actualResponse:   &Response{StatusCode: 200, Status: "200"},
-			expectedFilePath: "testdata/http_response_files/validator_partial_status_code_mismatch.hresp",
-			expectedErrCount: 0,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateResponses(tt.expectedFilePath, tt.actualResponse)
-
-			if tt.expectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				// Use direct assertion for single, specific error messages if tt.expectedErrText is set
-				// Otherwise, use the multierror helper.
-				if tt.expectedErrText != "" && tt.expectedErrCount == 1 && len(tt.expectedErrTexts) == 0 {
-					require.Error(t, err)
-					merr, ok := err.(*multierror.Error)
-					require.True(t, ok, "Expected a multierror.Error for single error case with expectedErrText")
-					require.Len(t, merr.Errors, 1)
-					assert.Contains(t, merr.Errors[0].Error(), tt.expectedErrText)
-				} else {
-					assertMultierrorContains(t, err, tt.expectedErrCount, tt.expectedErrTexts)
-				}
-			}
-		})
-	}
-}
-
-func TestValidateResponses_Headers(t *testing.T) {
-	tests := []struct {
-		name             string
-		actualResponse   *Response
-		expectedFilePath string // Changed from expectedFileContent
-		expectedErrCount int
-		expectedErrTexts []string
-	}{
-		{
-			name:             "matching headers",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}, "X-Request-Id": {"123"}}},
-			expectedFilePath: "testdata/http_response_files/validator_headerscontain_key_missing.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "mismatching header value",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"text/html"}}},
-			expectedFilePath: "testdata/http_response_files/validator_headerscontain_key_missing.hresp", // Expects application/json
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"expected value 'application/json' for header 'Content-Type' not found"},
-		},
-		{
-			name:             "missing expected header",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"X-Other": {"value"}}},
-			expectedFilePath: "testdata/http_response_files/validator_headers_missing_exp.hresp", // Expects X-Custom-Header
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"expected header 'X-Custom-Header' not found"}, // Adjusted error message
-		},
-		{
-			name:             "extra actual header (should be ignored)",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}, "X-Extra": {"ignored"}}},
-			expectedFilePath: "testdata/http_response_files/validator_headerscontain_key_missing.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "matching multi-value headers (order preserved)",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Accept": {"application/json", "text/xml"}}},
-			expectedFilePath: "testdata/http_response_files/validator_headers_multival_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "mismatching multi-value headers (different order)",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Accept": {"application/json", "text/xml"}}},
-			expectedFilePath: "testdata/http_response_files/validator_headers_multival_mismatch_order.hresp", // Expects xml then json
-			expectedErrCount: 0,                                                                              // Should now pass as order is not strictly enforced for all values
-			// expectedErrTexts: []string{"header 'Accept' value mismatch: expected [text/xml application/json], got [application/json text/xml]"},
-		},
-		{
-			name:             "mismatching multi-value headers (different value)",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Accept": {"application/json", "application/pdf"}}},
-			expectedFilePath: "testdata/http_response_files/validator_headers_multival_match.hresp", // Expects json then text/plain
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"expected value 'text/xml' for header 'Accept' not found"}, // Adjusted for current logic
-		},
-		{
-			name:             "subset of multi-value headers (actual has more values)",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Accept": {"application/json", "text/xml", "application/pdf"}}},
-			expectedFilePath: "testdata/http_response_files/validator_headers_multival_subset.hresp", // Expects only application/json
-			expectedErrCount: 0,                                                                      // Should now pass as expected ["application/json"] is found in actual
-			// expectedErrTexts: []string{"header 'Accept' value mismatch: expected [application/json], got [application/json text/xml application/pdf]"},
-		},
-		{
-			name:             "case-insensitive header key matching",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}},
-			expectedFilePath: "testdata/http_response_files/validator_headers_case_insensitive_match.hresp", // Expected file has "content-type"
-			expectedErrCount: 0,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateResponses(tt.expectedFilePath, tt.actualResponse)
-
-			if tt.expectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				assertMultierrorContains(t, err, tt.expectedErrCount, tt.expectedErrTexts)
-			}
-		})
-	}
-}
-
-func TestValidateResponses_Body_ExactMatch(t *testing.T) {
-	body1 := "Hello World"
-	body2 := "Hello Go"
-	tests := []struct {
-		name             string
-		actualResponse   *Response
-		expectedFilePath string // Changed from expectedFileContent
-		expectedErrCount int
-		expectedErrTexts []string // Changed from expectedErrText
-	}{
-		{
-			name:             "matching body",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", BodyString: "{\"key\":\"value\"}"},
-			expectedFilePath: "testdata/http_response_files/validator_body_exact_match_ok.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "mismatching body",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", BodyString: body2},
-			expectedFilePath: "testdata/http_response_files/validator_body_exact_match_ok.hresp", // file has `body1` equivalent
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch"}, // Diff will be part of the message
-		},
-		{
-			name:             "empty body in file, actual has content",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", BodyString: body1},
-			expectedFilePath: "testdata/http_response_files/validator_body_exact_no_body_exp.hresp", // Empty body
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch"},
-		},
-		{
-			name:             "empty body in file, actual also empty",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", BodyString: ""},
-			expectedFilePath: "testdata/http_response_files/validator_body_exact_no_body_exp.hresp", // Empty body
-			expectedErrCount: 0,
-		},
-		// Examples for new test cases based on your original structure:
-		// {
-		// 	name:                "file has body, actual has no body (nil body string ptr)",
-		// 	actualResponse:      &Response{StatusCode: 200, Status: "200 OK" /* BodyString is nil */},
-		// 	expectedFilePath:    "testdata/http_response_files/validator_body_exact_match_ok.hresp",
-		// 	expectedErrCount:    1,
-		// 	expectedErrTexts:    []string{"body mismatch"},
-		// },
-		// {
-		// 	name:                "file has no body, actual has body",
-		// 	actualResponse:      &Response{StatusCode: 200, Status: "200 OK", BodyString: body1},
-		// 	expectedFilePath:    "testdata/http_response_files/validator_body_exact_no_body_exp.hresp", // No body in file
-		// 	expectedErrCount:    1,
-		// 	expectedErrTexts:    []string{"body mismatch"},
-		// },
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateResponses(tt.expectedFilePath, tt.actualResponse)
-
-			if tt.expectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				assertMultierrorContains(t, err, tt.expectedErrCount, tt.expectedErrTexts)
-			}
-		})
-	}
-}
-
-// TestValidateResponses_BodyContains tests the BodyContains logic.
-// Since ParseExpectedResponseFile does not populate ExpectedResponse.BodyContains,
-// this test verifies that the BodyContains logic in ValidateResponses is benign
-// (doesn't cause errors) when the expected response comes from a file.
-func TestValidateResponses_BodyContains(t *testing.T) {
-	tests := []struct {
-		name           string
-		actualResponse *Response
-		// expectedFileContent will be simple, as it cannot express BodyContains - Now expectedFilePath
-		expectedFilePath string // Was expectedFileContent
-		expectedErrCount int    // Should be 0 if actual matches file's explicit parts
-		expectedErrTexts []string
-	}{
-		{
-			name:           "BodyContains logic is not triggered by file (positive case)",
-			actualResponse: &Response{StatusCode: 200, Status: "200 OK", BodyString: "Hello World Wide Web"},
-			// Exact match for body
-			expectedFilePath: "testdata/http_response_files/validator_bodycontains_positive.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:           "BodyContains logic not triggered, body mismatch handled by exact check",
-			actualResponse: &Response{StatusCode: 200, Status: "200 OK", BodyString: "Hello World"},
-			// File expects "Universe" - this will cause an exact body mismatch.
-			// The BodyContains part of ValidateResponses will not run due to empty expected.BodyContains.
-			expectedFilePath: "testdata/http_response_files/validator_bodycontains_mismatch.hresp",
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateResponses(tt.expectedFilePath, tt.actualResponse)
-
-			if tt.expectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				assertMultierrorContains(t, err, tt.expectedErrCount, tt.expectedErrTexts)
-			}
-		})
-	}
-}
-
-// TestValidateResponses_BodyNotContains is similar to BodyContains.
-// It verifies that BodyNotContains logic in ValidateResponses is benign
-// when the expected response comes from a file, as the file cannot specify BodyNotContains.
-func TestValidateResponses_BodyNotContains(t *testing.T) {
-	tests := []struct {
-		name             string
-		actualResponse   *Response
-		expectedFilePath string // Was expectedFileContent
-		expectedErrCount int
-		expectedErrTexts []string
-	}{
-		{
-			name:             "BodyNotContains logic is not triggered by file (positive case)",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", BodyString: "Hello World"},
-			expectedFilePath: "testdata/http_response_files/validator_bodynotcontains_exact_mismatch.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "BodyNotContains logic not triggered, actual contains something, file expects different exact body",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", BodyString: "Hello Universe"},
-			expectedFilePath: "testdata/http_response_files/validator_bodynotcontains_exact_mismatch.hresp", // Exact body mismatch
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch"},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateResponses(tt.expectedFilePath, tt.actualResponse)
-
-			if tt.expectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				assertMultierrorContains(t, err, tt.expectedErrCount, tt.expectedErrTexts)
-			}
-		})
-	}
-}
-
-func TestValidateResponses_WithSampleFile(t *testing.T) {
-	sampleFilePath := "testdata/http_response_files/sample1.http" // Use the actual file
-
-	parsedSampleExpected, err := parseExpectedResponseFile(sampleFilePath)
-	require.NoError(t, err)
-	require.Len(t, parsedSampleExpected, 1)
-	sampleExpectedStruct := parsedSampleExpected[0]
-
-	baseActual := &Response{
-		StatusCode: *sampleExpectedStruct.StatusCode,
-		Status:     *sampleExpectedStruct.Status,
-		Headers:    make(http.Header),
-		BodyString: *sampleExpectedStruct.Body,
-	}
-	for k, v := range sampleExpectedStruct.Headers { // Deep copy headers
-		baseActual.Headers[k] = append([]string{}, v...)
-	}
-
-	tests := []struct {
-		name                      string
-		actualModifier            func(actual *Response)
-		expectedFileSource        string
-		expectedFileContentString string // No longer used by the logic, but kept for structure if needed later
-		expectedErrCount          int
-		expectedErrTexts          []string
-	}{
-		{
-			name:               "perfect match with sample1.http",
-			actualModifier:     func(actual *Response) {},
-			expectedFileSource: sampleFilePath,
-			expectedErrCount:   0,
-		},
-		{
-			name: "status code mismatch",
-			actualModifier: func(actual *Response) {
-				actual.StatusCode = 500
-			},
-			expectedFileSource: sampleFilePath,
-			expectedErrCount:   1,
-			expectedErrTexts:   []string{"status code mismatch: expected 200, got 500"},
-		},
-		{
-			name: "status string mismatch",
-			actualModifier: func(actual *Response) {
-				actual.Status = "200 Everything is Fine"
-			},
-			expectedFileSource: sampleFilePath,
-			expectedErrCount:   1,
-			expectedErrTexts:   []string{"status string mismatch: expected '200 OK', got '200 Everything is Fine'"},
-		},
-		{
-			name: "header value mismatch for Content-Type",
-			actualModifier: func(actual *Response) {
-				actual.Headers.Set("Content-Type", "text/plain")
-			},
-			expectedFileSource: sampleFilePath,
-			expectedErrCount:   1,
-			expectedErrTexts:   []string{"expected value 'application/json; charset=utf-8' for header 'Content-Type' not found"},
-		},
-		{
-			name: "missing expected header Date",
-			actualModifier: func(actual *Response) {
-				actual.Headers.Del("Date")
-			},
-			expectedFileSource: sampleFilePath,
-			expectedErrCount:   1,
-			expectedErrTexts:   []string{"expected header 'Date' not found"},
-		},
-		{
-			name: "body mismatch",
-			actualModifier: func(actual *Response) {
-				actual.BodyString = "{\"message\": \"this is not the sample body\"}"
-			},
-			expectedFileSource: sampleFilePath,
-			expectedErrCount:   1,
-			expectedErrTexts:   []string{"body mismatch"},
-		},
-		{
-			name:               "BodyContains logic not triggered by exact file match (positive case)",
-			actualModifier:     func(actual *Response) {},
-			expectedFileSource: sampleFilePath,
-			expectedErrCount:   0,
-		},
-		{
-			name:               "BodyContains logic not triggered, exact body mismatch from file",
-			actualModifier:     func(actual *Response) {},
-			expectedFileSource: "testdata/http_response_files/validator_withsample_bodycontains_exactmismatch.hresp",
-			expectedErrCount:   1,
-			expectedErrTexts:   []string{"body mismatch"},
-		},
-		{
-			name:               "BodyNotContains logic not triggered by exact file match (positive case)",
-			actualModifier:     func(actual *Response) {},
-			expectedFileSource: sampleFilePath,
-			expectedErrCount:   0,
-		},
-		{
-			name:               "BodyNotContains logic not triggered, exact body mismatch from file (actual contains something unwanted by this hypothetical check)",
-			actualModifier:     func(actual *Response) { actual.BodyString = "{\"title\": \"delectus aut autem\"}" },
-			expectedFileSource: "testdata/http_response_files/validator_withsample_bodynotcontains_exactmismatch.hresp",
-			expectedErrCount:   1,
-			expectedErrTexts:   []string{"body mismatch"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			actualTest := &Response{
-				StatusCode: baseActual.StatusCode,
-				Status:     baseActual.Status,
-				Headers:    make(http.Header),
-				BodyString: baseActual.BodyString,
-			}
-			for k, v := range baseActual.Headers {
-				actualTest.Headers[k] = append([]string{}, v...)
-			}
-			tt.actualModifier(actualTest)
-
-			currentExpectedFilePath := tt.expectedFileSource
-
-			err := ValidateResponses(currentExpectedFilePath, actualTest)
-
-			if tt.expectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				assertMultierrorContains(t, err, tt.expectedErrCount, tt.expectedErrTexts)
-			}
-		})
-	}
-}
-
-// TODO: Add comprehensive tests combining multiple validation types
-// TODO: Add tests for JSONPath once that's part of ExpectedResponse and ValidateResponse
-// func TestValidateResponse_JSONPathChecks(t *testing.T) { ... } // Commented out for now
-// func TestValidateResponse_HeadersContain(t *testing.T) { ... } // Commented out for now
-
-// TestValidateResponses_HeadersContain verifies that HeadersContain logic in ValidateResponses
-// is benign when the expected response comes from a file, as the file cannot specify HeadersContain.
-func TestValidateResponses_HeadersContain(t *testing.T) {
-	tests := []struct {
-		name             string
-		actualResponse   *Response
-		expectedFilePath string // Was expectedFileContent
-		expectedErrCount int
-		expectedErrTexts []string
-	}{
-		{
-			name:             "HeadersContain logic not triggered (matching case for other fields)",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json; charset=utf-8"}}},
-			expectedFilePath: "testdata/http_response_files/validator_headerscontain_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "HeadersContain logic not triggered (mismatch on standard header)",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"text/html"}}},
-			expectedFilePath: "testdata/http_response_files/validator_headerscontain_key_missing.hresp", // Standard header mismatch
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"expected value 'application/json' for header 'Content-Type' not found"},
-		},
-		{
-			name:             "HeadersContain logic not triggered (expected header key not found by standard check)",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"X-Other": {"value"}}},
-			expectedFilePath: "testdata/http_response_files/validator_headerscontain_key_missing.hresp", // Expected header 'Content-Type' missing from actual
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"expected header 'Content-Type' not found"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateResponses(tt.expectedFilePath, tt.actualResponse)
-
-			if tt.expectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				assertMultierrorContains(t, err, tt.expectedErrCount, tt.expectedErrTexts)
-			}
-		})
-	}
-}
-
-func TestValidateResponses_PartialExpected(t *testing.T) {
-	tests := []struct {
-		name             string
-		actualResponse   *Response
-		expectedFilePath string // Was expectedFileContent
-		expectedErrCount int
-		expectedErrTexts []string
-	}{
-		{
-			name: "SCENARIO-LIB-009-005 Equiv: Expected file has only status code - match",
-			actualResponse: &Response{
-				StatusCode: 200,
-				Status:     "200",
-				Headers:    http.Header{"Content-Type": {"application/json"}},
-				BodyString: "",
-			},
-			expectedFilePath: "testdata/http_response_files/validator_partial_status_code_mismatch.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-009-005 Corrected: File has status code and empty body - actual matches",
-			actualResponse:   &Response{StatusCode: 200, Status: "200", BodyString: ""},
-			expectedFilePath: "testdata/http_response_files/validator_partial_status_code_mismatch.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-009-005-003 Corrected: File has status code and empty body - actual body mismatch",
-			actualResponse:   &Response{StatusCode: 200, Status: "200", BodyString: "non-empty body"},
-			expectedFilePath: "testdata/http_response_files/validator_partial_status_code_mismatch.hresp",
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch"},
-		},
-		{
-			name: "SCENARIO-LIB-009-005-004 Equiv: Expected file has only status code - status code mismatch",
-			actualResponse: &Response{
-				StatusCode: 404,
-				Status:     "404",
-				Headers:    http.Header{"Content-Type": {"application/json"}},
-				BodyString: "",
-			},
-			expectedFilePath: "testdata/http_response_files/validator_partial_status_code_mismatch.hresp",
-			expectedErrCount: 2,
-			expectedErrTexts: []string{"status code mismatch: expected 200, got 404", "status string mismatch: expected '200', got '404'"},
-		},
-		{
-			name: "SCENARIO-LIB-009-006 Equiv: Expected file has only specific headers (and status, empty body) - match",
-			actualResponse: &Response{
-				StatusCode: 200,
-				Status:     "200",
-				Headers:    http.Header{"Content-Type": {"application/json"}, "X-Custom": {"val"}},
-				BodyString: "",
-			},
-			expectedFilePath: "testdata/http_response_files/validator_partial_headers_key_missing.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name: "SCENARIO-LIB-009-006-002 Equiv: Expected file has only specific headers - header value mismatch",
-			actualResponse: &Response{
-				StatusCode: 200,
-				Status:     "200",
-				Headers:    http.Header{"Content-Type": {"text/plain"}, "X-Custom": {"val"}},
-				BodyString: "",
-			},
-			expectedFilePath: "testdata/http_response_files/validator_partial_headers_key_missing.hresp",
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"expected value 'application/json' for header 'Content-Type' not found"},
-		},
-		{
-			name: "SCENARIO-LIB-009-006-003 Equiv: Expected file has only specific headers - header missing in actual",
-			actualResponse: &Response{
-				StatusCode: 200,
-				Status:     "200",
-				Headers:    http.Header{"X-Other": {"value"}},
-				BodyString: "",
-			},
-			expectedFilePath: "testdata/http_response_files/validator_partial_headers_key_missing.hresp",
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"expected header 'Content-Type' not found"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateResponses(tt.expectedFilePath, tt.actualResponse)
-
-			if tt.expectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				assertMultierrorContains(t, err, tt.expectedErrCount, tt.expectedErrTexts)
-			}
-		})
-	}
-}
-
-func TestValidateResponses_StatusCode(t *testing.T) {
-	tests := []struct {
-		name                string
-		actualResponse      *Response
-		actualResponseCode  *int   // Pointer to allow nil
-		expectedFileContent string // To become FilePath
-		expectedFilePath    string // New
-		expectedErrCount    int
-		expectedErrText     string
-	}{
-		{
-			name:               "matching status code",
-			actualResponseCode: ptr(200),
-			expectedFilePath:   "testdata/http_response_files/validator_partial_status_code_mismatch.hresp",
-			expectedErrCount:   0,
-		},
-		{
-			name:               "mismatching status code",
-			actualResponseCode: ptr(404),
-			expectedFilePath:   "testdata/http_response_files/validator_partial_status_code_mismatch.hresp", // Expect 200
-			expectedErrCount:   2,                                                                           // Expects 2 errors: status code and status string mismatch
-			expectedErrText:    "",                                                                          // Clear this as we use currentExpectedErrTexts now
-			// currentExpectedErrTexts will be: ["status code mismatch: expected 200, got 404", "status string mismatch: expected '200', got '404'"]
-		},
-		{
-			name:               "nil actual status code (should not happen with real http.Response)",
-			actualResponseCode: nil, // Actual code is 0, so if file expects 200, it's a mismatch
-			expectedFilePath:   "testdata/http_response_files/validator_partial_status_code_mismatch.hresp",
-			expectedErrCount:   2,  // Expects 2 errors: status code and status string mismatch
-			expectedErrText:    "", // Clear this
-			// currentExpectedErrTexts will be: ["status code mismatch: expected 200, got 0", "status string mismatch: expected '200', got '0'"]
-		},
-		{
-			name:               "nil expected status code (file has no status line)",
-			actualResponseCode: ptr(200),
-			expectedFilePath:   "testdata/http_response_files/validator_status_code_nil_expected.hresp",
-			expectedErrCount:   2, // This will fail parsing + count mismatch
-			// currentExpectedErrTexts will be: ["failed to parse expected response file", "mismatch in number of responses"]
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			actual := &Response{}
-			if tt.actualResponseCode != nil {
-				actual.StatusCode = *tt.actualResponseCode
-				actual.Status = fmt.Sprintf("%d", actual.StatusCode) // Populate actual.Status
-			} else {
-				actual.Status = "0"   // Default if StatusCode is 0 (from nil actualResponseCode)
-				actual.StatusCode = 0 // ensure StatusCode is also 0 if actualResponseCode is nil
-			}
-
-			currentExpectedErrCount := tt.expectedErrCount
-			var currentExpectedErrTexts []string
-
-			if tt.name == "mismatching status code" {
-				currentExpectedErrTexts = []string{"status code mismatch: expected 200, got 404", "status string mismatch: expected '200', got '404'"}
-			} else if tt.name == "nil actual status code (should not happen with real http.Response)" {
-				currentExpectedErrTexts = []string{"status code mismatch: expected 200, got 0", "status string mismatch: expected '200', got '0'"}
-			} else if tt.name == "nil expected status code (file has no status line)" {
-				currentExpectedErrTexts = []string{
-					"failed to parse expected response file", // This will contain the "invalid status code 'application/json'" detail
-					"mismatch in number of responses: got 1 actual, but expected 0",
-				}
-			} else if tt.expectedErrText != "" {
-				currentExpectedErrTexts = []string{tt.expectedErrText}
-			}
-
-			err := ValidateResponses(tt.expectedFilePath, actual)
-
-			if currentExpectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				require.Error(t, err)
-				merr, ok := err.(*multierror.Error)
-				require.True(t, ok, "Expected a multierror.Error")
-				assert.Len(t, merr.Errors, currentExpectedErrCount)
-				// Check specific error texts if provided for the adjusted expectation
-				if len(currentExpectedErrTexts) > 0 {
-					for _, expectedText := range currentExpectedErrTexts {
-						found := false
-						for _, e := range merr.Errors {
-							if strings.Contains(e.Error(), expectedText) {
-								found = true
-								break
-							}
-						}
-						assert.True(t, found, "Expected error text '%s' not found in %v", expectedText, merr.Errors)
-					}
-				} else if tt.expectedErrText != "" && currentExpectedErrCount == 1 { // Original single error check
-					assert.ErrorContains(t, merr.Errors[0], tt.expectedErrText)
-				}
-			}
-		})
-	}
-}
-
-func TestValidateResponses_BodyRegexpPlaceholder(t *testing.T) {
-	tests := []struct {
-		name             string
-		actualResponse   *Response
-		expectedFilePath string
-		expectedErrCount int
-		expectedErrTexts []string
-	}{
-		{
-			name:             "SCENARIO-LIB-022-001: simple regexp match",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"id": "123"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_regexp_simple_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-022-002: simple regexp no match",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"status": "FAILED"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_regexp_simple_no_match.hresp",
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch (regexp/placeholder evaluation failed)", "Compiled Regex: ^\\{\"status\": \"(SUCCESS)\"\\}$"},
-		},
-		{
-			name:             "SCENARIO-LIB-022-003: multiple regexp match",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"userId": "U-abc", "transactionId": "T-12345"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_regexp_multiple_match.hresp",
-			expectedErrCount: 0,
-		},
-		// {
-		// 	name:             "SCENARIO-LIB-022-004: regexp with special characters in pattern",
-		// 	actualResponse:   &Response{StatusCode: 200, Status: "200 OK", BodyString: `Value: 123.test`},
-		// 	expectedFilePath: "testdata/http_response_files/validator_body_regexp_special_chars.hresp",
-		// 	expectedErrCount: 0,
-		// },
-		{
-			name:             "SCENARIO-LIB-022-005: invalid regexp pattern",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", BodyString: `{"data": "value"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_regexp_invalid_pattern.hresp",
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"failed to compile master regex from expected body", "error parsing regexp"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateResponses(tt.expectedFilePath, tt.actualResponse)
-			if tt.expectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				assertMultierrorContains(t, err, tt.expectedErrCount, tt.expectedErrTexts)
-			}
-		})
-	}
-}
-
-func TestValidateResponses_BodyAnyGuidPlaceholder(t *testing.T) {
-	tests := []struct {
-		name             string
-		actualResponse   *Response
-		expectedFilePath string
-		expectedErrCount int
-		expectedErrTexts []string
-	}{
-		{
-			name:             "SCENARIO-LIB-023-001: valid GUID match",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"correlationId": "123e4567-e89b-12d3-a456-426614174000"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anyguid_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-023-002: not a GUID",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"id": "not-a-guid"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anyguid_no_match.hresp",
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch (regexp/placeholder evaluation failed)"},
-		},
-		{
-			name:             "SCENARIO-LIB-023-003: GUID in larger text",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", BodyString: `Session started with ID: 123e4567-e89b-12d3-a456-426614174000. Please use this for subsequent requests.`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anyguid_larger_text.hresp",
-			expectedErrCount: 0,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateResponses(tt.expectedFilePath, tt.actualResponse)
-			if tt.expectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				assertMultierrorContains(t, err, tt.expectedErrCount, tt.expectedErrTexts)
-			}
-		})
-	}
-}
-
-func TestValidateResponses_BodyAnyTimestampPlaceholder(t *testing.T) {
-	tests := []struct {
-		name             string
-		actualResponse   *Response
-		expectedFilePath string
-		expectedErrCount int
-		expectedErrTexts []string
-	}{
-		{
-			name:             "SCENARIO-LIB-024-001: valid integer timestamp match",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"createdAt": "1678886400"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anytimestamp_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-024-002: not an integer timestamp (string)",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"timestamp": "not-a-timestamp"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anytimestamp_no_match.hresp",
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch (regexp/placeholder evaluation failed)"},
-		},
-		{
-			name:             "SCENARIO-LIB-024-003: not an integer timestamp (float)",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"eventTime": "1678886400.5"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anytimestamp_float.hresp",
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch (regexp/placeholder evaluation failed)"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateResponses(tt.expectedFilePath, tt.actualResponse)
-			if tt.expectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				assertMultierrorContains(t, err, tt.expectedErrCount, tt.expectedErrTexts)
-			}
-		})
-	}
-}
-
-func TestValidateResponses_BodyAnyDatetimePlaceholder(t *testing.T) {
-	tests := []struct {
-		name             string
-		actualResponse   *Response
-		expectedFilePath string
-		expectedErrCount int
-		expectedErrTexts []string
-	}{
-		{
-			name:             "SCENARIO-LIB-025-001: rfc1123 match",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"lastModified": "Tue, 15 Mar 2023 12:00:00 GMT"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anydatetime_rfc1123_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-025-002: iso8601 match (RFC3339)",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"eventTime": "2023-03-15T12:00:00Z"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anydatetime_iso8601_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-025-002b: iso8601 match with offset",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"eventTime": "2023-03-15T12:00:00+01:00"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anydatetime_iso8601_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-025-002c: iso8601 match with milliseconds and Z",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"eventTime": "2023-03-15T12:00:00.123Z"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anydatetime_iso8601_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-025-003: custom Go layout match",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"date": "2023-03-15"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anydatetime_custom_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-025-004: format mismatch (rfc1123 expected, actual is YYYY-MM-DD)",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"timestamp": "2023-03-15"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anydatetime_format_mismatch.hresp",
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch (regexp/placeholder evaluation failed)"},
-		},
-		{
-			name:             "SCENARIO-LIB-025-005: invalid format keyword",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"time": "12:34:56"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anydatetime_invalid_keyword.hresp",
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch (regexp/placeholder evaluation failed)", "(\\z.\\A)"},
-		},
-		{
-			name:             "SCENARIO-LIB-025-006: missing format argument",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"time": "12:34:56"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anydatetime_missing_format.hresp",
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch (regexp/placeholder evaluation failed)", "(\\z.\\A)"},
-		},
-		{
-			name:             "custom format empty string literal \"\" - should fail",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"date": ""}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_anydatetime_custom_empty_format.hresp",
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch (regexp/placeholder evaluation failed)", "(\\z.\\A)"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateResponses(tt.expectedFilePath, tt.actualResponse)
-			if tt.expectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				assertMultierrorContains(t, err, tt.expectedErrCount, tt.expectedErrTexts)
-			}
-		})
-	}
-}
-
-func TestValidateResponses_BodyAnyPlaceholder(t *testing.T) {
-	tests := []struct {
-		name             string
-		actualResponse   *Response
-		expectedFilePath string
-		expectedErrCount int
-		expectedErrTexts []string
-	}{
-		{
-			name:             "SCENARIO-LIB-026-001: $any matching a simple string",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"key": "some value"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_any_simple_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-026-002: $any matching special characters and spaces",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"text/plain"}}, BodyString: `Value: !@#$%^&*()_+{}[];':\",./<>?         end`},
-			expectedFilePath: "testdata/http_response_files/validator_body_any_special_chars_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-026-003: $any matching an empty string segment",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"prefix": "", "data": "", "suffix": ""}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_any_empty_segment_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-026-004: $any matching a multi-line string",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"text/plain"}}, BodyString: "Start:\nThis is line 1.\nThis is line 2.\nAnd line 3.\nEnd."},
-			expectedFilePath: "testdata/http_response_files/validator_body_any_multiline_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "SCENARIO-LIB-026-005: multiple $any placeholders",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"field1": "value1", "field2": "constant", "field3": "value3"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_any_multiple_placeholders_match.hresp",
-			expectedErrCount: 0,
-		},
-		{
-			name:             "$any fails if preceding literal doesn't match",
-			actualResponse:   &Response{StatusCode: 200, Status: "200 OK", Headers: http.Header{"Content-Type": {"application/json"}}, BodyString: `{"wrong_key": "some value"}`},
-			expectedFilePath: "testdata/http_response_files/validator_body_any_simple_match.hresp", // Expects {"key": ...}
-			expectedErrCount: 1,
-			expectedErrTexts: []string{"body mismatch"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			err := ValidateResponses(tt.expectedFilePath, tt.actualResponse)
-			if tt.expectedErrCount == 0 {
-				assert.NoError(t, err)
-			} else {
-				assertMultierrorContains(t, err, tt.expectedErrCount, tt.expectedErrTexts)
-			}
-		})
-	}
-}
